pyproject.toml
[project]
name = "mitre-emb3d"
version = "0.4.0"
description = "MITRE EMB3D: Library & CLI for working with the MITRE EMB3D framework."
readme = "README.md"
authors = [
    { name = "Sachdeva, Kapil"}
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.1.0",
    "google-adk>=1.27.1",
    "httpx>=0.28.1",
    "litellm>=1.82.2",
    "networkx>=3.6.1",
    "pydantic>=2.12.5",
    "textual>=8.0.2",
    "typer>=0.16.0",
    "xdg-base-dirs>=6.0.2",
]

classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: Apache Software License",
  "Typing :: Typed"
]

keywords = [
    "MITRE",
    "MITRE EMB3D",
    "security",
    "threat modeling"
]

[project.urls]
repository = "https://github.com/ksachdeva/mitre-emb3d"
issues = "https://github.com/ksachdeva/mitre-emb3d/issues"

[project.scripts]
mitre-emb3d = "mitre_emb3d.__main__:cli_app"
med = "mitre_emb3d.__main__:cli_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.6.1",
    "pytest-mock>=3.14.0",
    "poethepoet>=0.30.0",
    "packaging>=24.2",
    "pyright>=1.1.378",
    "mypy>=1.10.0",
    "ruff>=0.4.8",
    "pre-commit>=4.3.0",
    "textual-dev>=1.8.0",
]

[tool.ruff]
line-length = 120
fix = true
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501"]

[tool.ruff.lint.flake8-tidy-imports]
[tool.ruff.lint.flake8-tidy-imports.banned-api]
"unittest".msg = "Use `pytest` instead."

[tool.ruff.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query"]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true

# from https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = false

disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false

[tool.pyright]
include = ["src", "tests", "samples"]
# typeCheckingMode = "strict"
reportUnnecessaryIsInstance = false
reportMissingTypeStubs = false
exclude = []

[tool.poe]
include = "poe.toml"
